Split out color utils + added tests
#1542
